+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-# Load My Chunks Version 1.0.5-alpha2
+# Load My Chunks Version 1.0.5-alpha3
 
 # WARNING THIS IS A ALPHA BUILD - THINGS MAY NOT WORK
 Alpha is only available for 1.21 and will be backported on release.
 
 ## Load My Computers - Computercraft Integration
 
-Patched some issues I caused with computer craft turtles due to a bad mixin.
-
+Patched an issue in LevelChunkMixin causing Block Entities to double tick occasionally. 
 Made the mod's chunk data thread safe, to fix a relevant conflict on *some* systems with mods like C^2ME which thread chunk loading.
+Made CC:T a completely soft dependency
