+++ gradle.properties
@@ -44,6 +44,3 @@ deps.mod.cct=[VERSIONED]
 publish.modrinth=OtbeKtn3
 publish.curseforge=1024152
 publish.alpha = true
-
-vcs=1.21
-vcs-env=fabric
